feat: almost full play vs bot

+++ api/app/model/game.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Integer, Boolean, ForeignKey, String
+from sqlalchemy import Column, DateTime, Integer, Boolean, ForeignKey, SmallInteger, String
 from sqlalchemy.orm import relationship
 from api.app.model.base import BareBaseModel
 
@@ -7,7 +7,7 @@ class Game(BareBaseModel):
 
     variant_id = Column(Integer, ForeignKey('variants.id'), nullable=False)
     created_at = Column(DateTime)
-    time_mode = Column(Integer)
+    time_mode = Column(SmallInteger)
     status = Column(Boolean, nullable=False)
     result = Column(Integer, nullable=False)
     slug = Column(String(8), nullable=False)
